threejs-package {threejs}	R Documentation

Interactive 3D graphics including point clouds and globes using three.js and htmlwidgets.

Description

Interactive 3D graphics including point clouds and globes using three.js and htmlwidgets.

References

http://threejs.org

Examples

## Not run: 
library("shiny")
runApp(system.file("examples/globe",package="threejs"))
runApp(system.file("examples/scatterplot",package="threejs"))

# See also help for globe.js and scatterplot3.js

## End(Not run)

[Package threejs version 0.3.1 Index]