globeOutput {threejs}	R Documentation

Shiny bindings for threejs widgets

Description

Output and render functions for using threejs widgets within Shiny applications and interactive Rmd documents.

Usage

globeOutput(outputId, width = "100%", height = "600px")

renderGlobe(expr, env = parent.frame(), quoted = FALSE)

scatterplotThreeOutput(outputId, width = "100%", height = "500px")

renderScatterplotThree(expr, env = parent.frame(), quoted = FALSE)

Arguments

outputId
output variable to read from

width, height
Must be a valid CSS unit (like "100%", "400px", "auto") or a number, which will be coerced to a string and have "px" appended.

expr
An expression that generates threejs graphics.

env
The environment in which to evaluate expr.

quoted
Is expr a quoted expression (with quote())? This is useful if you want to save an expression in a variable.

[Package threejs version 0.3.1 Index]